\name{PracticaTwitter-package}
\alias{PracticaTwitter-package}
\alias{PracticaTwitter}
\docType{package}
\title{
\packageTitle{PracticaTwitter}
}
\description{
\packageDescription{PracticaTwitter}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{PracticaTwitter}
\packageIndices{PracticaTwitter}
~~ An overview of how to use ~~
~~ the package, including the ~~
~~ most important functions ~~
}
\author{
\packageAuthor{PracticaTwitter}

Maintainer: \packageMaintainer{PracticaTwitter}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard ~~
~~ keywords, one per line, from ~~
~~ file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
